requirements.txt
allure-pytest==2.5.2
pytest==3.9.1
pytest-rerunfailures==4.2
selenium==3.14.1
webdriver-manager==1.7
pytest-html==1.19.0